+++ src/main/java/com/spring/javaGroupS7/dao/AdminDAO.java
@@ -31,6 +31,6 @@ public interface AdminDAO {
 
 	List<ProductOrderVO> getProductCnt();
 
-	int getMonthlySales();
+	long getMonthlySales();
 
 }
